+++ src/FSharp.Expandable.Compiler/Program.cs
@@ -20,7 +20,6 @@
 //////////////////////////////////////////////////////////////////////////////
 
 using System;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Reflection;
@@ -32,6 +31,42 @@ namespace FSharp.Expandable
     /// </summary>
     internal static class Program
     {
+        private static double ParseVersion(string version)
+        {
+            double value;
+            if (double.TryParse(version, out value))
+            {
+                return value;
+            }
+            else
+            {
+                return -1;
+            }
+        }
+
+        private static int? TryLoadAndRun(string corePath, string[] args)
+        {
+            MethodInfo mi = null;
+
+            try
+            {
+                var assembly = Assembly.LoadFrom(corePath);
+                var type = assembly.GetType("FSharp.Expandable.Compiler");
+                mi = type.GetMethod("DefaultDriver");
+            }
+            catch
+            {
+                return null;
+            }
+
+            if (mi == null)
+            {
+                return null;
+            }
+
+            return (int)mi.Invoke(null, new object[] {args});
+        }
+
         /// <summary>
         /// Main entry point.
         /// </summary>
@@ -39,64 +74,52 @@ namespace FSharp.Expandable
         /// <returns>Return value</returns>
         public static int Main(string[] args)
         {
+            // STRATEGY:
+            //   Completely late bound for another package placed FSharp.Expandable.Compiler.Core.dll.
+            //   Because Visual Studio loaded and locked FSharp.Expandable.Compiler.Tasks.dll,
+            //   fscx infrastructure and filters FRUSTRATED debugging many recycled and recycled VS process...
+
             ///////////////////////////////////////////////////////////////////////
-            // Extract arguments.
+            // Search FSharp.Expandable.Compiler.Core.dll
 
-            var arguments = args.ExtractCompilerArguments();
+            // NuGet package structure:
+            //   packages --+-- FSharp.Expandable.Compiler.Build --+-- build --+-- FSharp.Expandable.Compiler.Tasks.dll (1: from MSBuild)
+            //              |                                                  +-- fscx.exe (2: invoke from FscTask)
+            //              +-- FSharp.Expandable.Compiler.Core --+-- lib --+-- net45 --+-- FSharp.Expandable.Compiler.Core.dll (3: DefaultDriver() function)
 
-            // If not giving visitor paths:
-            if (arguments.VisitorPaths.Any() == false)
+            var exeLocation = new Uri(typeof(Program).Assembly.CodeBase).LocalPath;
+            var packagesPath = Path.Combine(Path.GetDirectoryName(exeLocation), "..", "..");
+
+            var searchFolderBases =
+              Directory.Exists(packagesPath)
+              ? from packagePath in
+                  Directory.EnumerateDirectories(packagesPath, "*", SearchOption.TopDirectoryOnly)
+                let packageName = Path.GetFileName(packagePath)
+                where packageName.StartsWith("FSharp.Expandable.Compiler.Core.")
+                let version = ParseVersion(packageName.Substring("FSharp.Expandable.Compiler.Core.".Length))
+                orderby version descending
+                from libPath in
+                  Directory.EnumerateDirectories(packagePath, "lib", SearchOption.TopDirectoryOnly)
+                from monikerPath in
+                  Directory.EnumerateDirectories(packagePath, "net45", SearchOption.TopDirectoryOnly)
+                select monikerPath
+              : new[] { "." };
+
+            var corePaths =
+                from searchFolderBase in searchFolderBases
+                from dllPath in Directory.EnumerateFiles(searchFolderBase, "*.dll", SearchOption.AllDirectories)
+                select dllPath;
+
+            foreach (var corePath in corePaths)
             {
-                ///////////////////////////////////////////////////////////////////////
-                // Crawl visitor assemblies
-
-                // NuGet package structure:
-                //   packages --+-- fscx-0.1.*     --+-- build --+-- fscx.exe
-                //              +-- HogeFilter-1.0 --+-- build --+-- HogeFilter.dll
-                //              +-- HagaFilter-1.0 --+-- build --+-- HagaFilter.dll
-
-                var exeLocation = new Uri(Assembly.GetExecutingAssembly().CodeBase).LocalPath;
-                var packagesPath = Path.Combine(Path.GetDirectoryName(exeLocation), "..", "..");
-
-                var searchFolderBases =
-                    Directory.Exists(packagesPath)
-                        ? from packagePath in
-                            Directory.EnumerateDirectories(packagesPath, "*", SearchOption.TopDirectoryOnly)
-                          from buildPath in
-                            Directory.EnumerateDirectories(packagePath, "build", SearchOption.TopDirectoryOnly)
-                          select buildPath
-                        : new[] { "." };
-
-                var visitorPaths =
-                    (from searchFolderBase in
-                        searchFolderBases
-                     from dllPath in
-                        Directory.EnumerateFiles(searchFolderBase, "*.dll", SearchOption.AllDirectories)
-                     select dllPath).
-                    FilterVisitors();
-#if DEBUG
-                foreach (var path in visitorPaths)
+                var result = TryLoadAndRun(corePath, args);
+                if (result.HasValue)
                 {
-                    Debug.WriteLine(path);
+                    return result.Value;
                 }
-#endif
-                // Place visitor paths.
-                arguments.VisitorPaths = visitorPaths;
             }
 
-            ///////////////////////////////////////////////////////////////////////
-            // Compile
-
-            return Compiler.CompileWithArguments(logEntry =>
-               Console.WriteLine(
-                   "{0}({1},{2}): {3}{4}: {5}",
-                   logEntry.FileName,
-                   logEntry.Line,
-                   logEntry.Column,
-                   logEntry.Type.ToString().ToLowerInvariant(),
-                   string.IsNullOrWhiteSpace(logEntry.Code) ? "" : (" " + logEntry.Code),
-                   logEntry.Message),
-               arguments);
+            throw new FileNotFoundException("Cannot found FSharp.Expandable.Compiler.Core.dll");
         }
     }
 }